KL-159/rename: change directory
