+++ src/components/ProductList/ProductList.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import PreviewContent from '../../../../components/PreviewContent/PreviewContent'
-import { user } from '../../../../test/data/DummyData'
+import PreviewContent from '../PreviewContent/PreviewContent'
+import { user } from '../../test/data/DummyData'
 import StyledList from './ProductList.style'
 
 function ProductList({ dataList }) {
