+++ src/components/ProductList/ProductList.style.js
@@ -1,15 +1,15 @@
 import styled from 'styled-components'
-import theme from '../../../../styles/theme'
+import theme from '../../styles/theme'
 
 const StyledList = styled.article`
   min-width: 36rem;
   margin-top: 0.5rem;
-  margin-bottom: 1.5rem;
+  margin-bottom: 3.125rem;
   display: grid;
   grid-template-columns: repeat(3, minmax(auto, 11rem));
   gap: 1.4rem 1.5rem;
   &.empty {
-    font-color: ${theme.color.textGrey};
+    color: ${theme.color.textGrey};
     min-height: 15rem;
     margin-bottom: 0;
     display: flex;
